fix: mutate PodInfo in place on update instead of replace-and-copy
Refactors AddOrUpdatePod to update the existing PodInfo struct in place
rather than creating a new one and copying metrics. This eliminates a
race window between releasing the read lock and storing the new struct,
and aligns with how the rest of the codebase mutates PodInfo.

Also merges tests into store_test.go, renames misleading test case,
and runs gofmt.
